requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
async-timeout>=4.0.2
yt-dlp>=2023.3.4
python-ffmpeg>=2.0.2
aiohttp>=3.8.5
requests>=2.28.2
aiohappyeyeballs>=2.6.0
aiosignal>=1.3.0
attrs>=25.2.0
frozenlist>=1.4.0
multidict>=6.0.0
yarl>=1.9.0
PyNaCl>=1.5.0 